requirements.txt
psycopg2==2.9.3
psycopg2-binary==2.9.3
requests==2.28.1
gspread==5.5.0
oauth2client==4.1.3
google-api-python-client==2.60.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
telebot==0.0.4
Telethon==1.25.0
Django==4.1.1
django-google-charts==0.1.1
django-qsstats-magic==1.1.0
django-tables2==2.4.1
pandas==1.4.4
